Add slot structure and functions to draw them
Implement a slot structure ("slot.h") for use with the game board.
Slots occupy an (x, y) position of the screen, have a width and a
height, and hold a numerical value. These slots are used for
displaying the game board and providing the user with a way to
interact with it.

Additionally, move the image assets into their own folder (named
"assets") and move the constants to their own file ("constants.c").

Begin work on methods to track and respond to mouse movement.