Complete core functions
Finalize the core procedures of the game by:
	-Adding functions ("matrix.h") to control the game board
	-Expanding the slot structure to interact with the game board
	-Isolating drawing code from slot structure into "slot_draw.h"
	-Completing the game logic in main()
	-Moving functions not related to a structure into "functions.h"

Also, remove image assets from starting directory, as they have
previously been placed into the "assets" folder.

The game is fully playable with two human players, although it is not
user-friendly. Window must be closed and reopened at the end of
each match to continue playing. Some minor drawing bugs exist, but the
game is stable.

Going forward, additional code changes should begin in their own branch
and merged when completed. The master branch should be reserved for
stable and playable updates to the game.